Revert "Routine dev merge to master (#69)"
This reverts commit c03030d634c6959f808168e14292508ad950707f.
