+++ controllers/userController.js
@@ -1,7 +1,11 @@
+const { generateToken } = require("../middlewares/auth");
+const { generateUsername } = require("unique-username-generator");
 const User = require("../models/userModel");
+const ErrorHandler = require("../utils/errorHandler");
 const sendToken = require("../utils/jwtToken");
+const crypto = require("crypto");
+const { ObjectId } = require("mongodb");
 const { activityLogger, errorLogger } = require("../utils/logger");
-const bcrypt = require("bcryptjs");
 const {
   CITY_TO_COORDINATE,
   AVAILABLE_CITIES,
@@ -47,21 +51,8 @@ exports.updateLocation = async (req, res, next) => {
   try {
     const userLocation = body?.userLocation;
     const cityLocation = body?.cityLocation;
-    const homeLocation = body?.homeLocation;
     let updatedCoordinates;
-    if (homeLocation) {
-      // If userLocation is provided, it should be an array with [lat, lng]
-      activityLogger.info(
-        "Updating user's home location based on coordinates: " + homeLocation
-      );
-      updatedCoordinates = homeLocation;
-      await User.findByIdAndUpdate(user._id, {
-        $set: {
-          "home_coordinates.coordinates": homeLocation,
-        },
-      });
-    }
-    else if (userLocation) {
+    if (userLocation) {
       // If userLocation is provided, it should be an array with [lat, lng]
       activityLogger.info(
         "Updating user's location based on coordinates: " + userLocation
@@ -70,7 +61,7 @@ exports.updateLocation = async (req, res, next) => {
       await User.findByIdAndUpdate(user._id, {
         $set: {
           "current_coordinates.coordinates": userLocation,
-          "city.coordinates": [0,0]
+          "city.coordinates": [0,0],
         },
       });
     } else if (cityLocation && CITY_TO_COORDINATE[cityLocation.toLowerCase()]) {
@@ -83,7 +74,7 @@ exports.updateLocation = async (req, res, next) => {
       await User.findByIdAndUpdate(user._id, {
         $set: {
           "city.coordinates": coordinates,
-          "current_coordinates.coordinates": [0,0]
+          "current_coordinates.coordinates": [0,0],
         },
       });
     } else {
@@ -134,6 +125,80 @@ exports.getUserGroups = async (req, res, next) => {
       `Error in getUserGroups for ${user.username}. Error: ${error}`
     );
   }
+};
+
+// User Login
+exports.loginUser = async (req, res, next) => {
+  const { userId, password } = req.body;
+  let email = "";
+  let username = "";
+  let user;
+
+  const emailRegex = /^[\w-]+(\.[\w-]+)*@([\w-]+\.)+[a-zA-Z]{2,7}$/;
+
+  if (emailRegex.test(userId)) {
+    email = userId;
+    user = await User.findOne({ email: email });
+  } else {
+    username = userId;
+    user = await User.findOne({ username: username });
+  }
+
+  if (!user) {
+    return next(new ErrorHandler("Invalid Email or Password", 401));
+  }
+
+  const match = await user.comparePassword(password);
+
+  if (!match) {
+    errorLogger.error("An unexpected error occurred during login:", error);
+    return next(new ErrorHandler("Invalid Email or Password", 401));
+  }
+  activityLogger.info(
+    `User ${user.username}(${user._id}) has logged in successfully`
+  );
+  sendToken(user, 200, res);
+};
+
+// User Register
+exports.registerUser = async (req, res) => {
+  const { password, email, pic, current_coordinates } = req.body;
+  let username = generateUsername() + Math.floor(Math.random() * 10000);
+  while (await User.findOne({ username })) {
+    username = generateUsername() + Math.floor(Math.random() * 10000);
+    activityLogger.info(
+      `Registration attempt for user with username ${username}.`
+    );
+  }
+  try {
+    const picture = `https://api.multiavatar.com/${username}.png`;
+    const user = await User.create({
+      username: username,
+      password: password,
+      email: email,
+      picture: picture,
+    });
+
+    sendToken(user, 200, res);
+  } catch (error) {
+    errorLogger.error(
+      "An unexpected error occurred during user registration:",
+      error
+    );
+    if (error.code === 11000 || error.code === 11001) {
+      return res.status(400).json({
+        error: "Duplicate Entry",
+        message: Object.keys(error.keyValue)[0] + " already exists.",
+      });
+    }
+    return res.status(400).json(error);
+  }
+};
+
+//Logout User
+exports.logoutUser = async (req, res, next) => {
+  res.clearCookie("token");
+  res.end();
 };
 
 // update user display pictures
@@ -198,51 +263,6 @@ exports.deleteUser = async (req, res) => {
     const newData = await User.deleteOne({ _id: user._id });
     res.status(200).json(newData);
   } catch (error) {
-    errorLogger.error(`An error occured :`, error);
     res.status(500);
   }
 };
-
-exports.changePassword = async (req, res) => {
-  const { currentPassword, newPassword } = req.body;
-  const user = await User.findById({ _id: req.user._id });
-  const match = await user.comparePassword(currentPassword);
-  try {
-    if (match) {
-      const encryptPassword = await bcrypt.hash(newPassword, 10);
-      const update = await User.updateOne(
-        { _id: user._id },
-        { $set: { password: encryptPassword } }
-      );
-      res.status(200).json(update);
-    }
-    else {
-      errorLogger.error("Wrong password while changing password");
-      res.status(401).json({
-        msg: "wrong current password"
-      });
-    }
-  } catch (err) {
-    errorLogger.error(
-      "An unexpected error occurred during change Password:",
-      err
-    );
-  }
-}
-
-exports.findMe = async (req, res) => {
-  const user = req.user;
-  try {
-    const findme = await User.updateOne(
-      { _id: user._id },
-      { $set: { findMe: !user.findMe } }
-    );
-    activityLogger.info('Updated find me option as', !user.findMe)
-    res.status(200).json(findme);
-  } catch (err) {
-    errorLogger.error('There is an error in findMe API: ', err);
-    res.status(500).json({
-      msg: "Find me API crashed"
-    });
-  }
-}
