+++ models/userModel.js
@@ -10,6 +10,7 @@ const userSchema = new mongoose.Schema({
   },
   password: {
     type: String,
+    required: [true, "Please enter password"],
   },
   picture: {
     type: String,
@@ -27,18 +28,7 @@ const userSchema = new mongoose.Schema({
     },
     coordinates: {
       type: [Number],
-      default: [0, 0],
-    },
-  },
-  home_coordinates: {
-    type: {
-      type: String,
-      enum: ["Point"],
-      default: "Point",
-    },
-    coordinates: {
-      type: [Number],
-      default: [0, 0],
+      default: [0,0],
     },
   },
   city: {
@@ -49,7 +39,7 @@ const userSchema = new mongoose.Schema({
     },
     coordinates: {
       type: [Number],
-      default: [0, 0],
+      default: [0,0],
     },
   },
   findMe: {
@@ -66,14 +56,10 @@ const userSchema = new mongoose.Schema({
     type: Number,
     default: 1000,
   },
-  auth_type: {
-    type: String
-  }
 });
 
 userSchema.pre("save", async function () {
-  if (this.password !== undefined)
-    this.password = await bcrypt.hash(this.password, 10);
+  this.password = await bcrypt.hash(this.password, 10);
 });
 
 userSchema.methods.getJWTToken = function () {
