+++ package.json
@@ -11,28 +11,19 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "@multiavatar/multiavatar": "^1.0.7",
     "aws-sdk": "^2.1596.0",
     "bcryptjs": "^2.4.3",
     "cookie-parser": "^1.4.6",
     "cors": "^2.8.5",
     "dotenv": "^16.3.1",
     "express": "^4.18.2",
-    "express-session": "^1.18.0",
     "faker": "^5.5.3",
     "jsonwebtoken": "^9.0.2",
     "mongoose": "^8.0.0",
     "node-file-logger": "^0.9.5",
     "nodemon": "^3.0.1",
     "opencage-api-client": "^1.0.3",
-    "otp-generator": "^4.0.1",
-    "passport": "^0.7.0",
-    "passport-google-oauth2": "^0.2.0",
-    "pg": "^8.11.5",
-    "pg-hstore": "^2.3.4",
-    "sequelize": "^6.37.3",
     "socket.io": "^4.7.2",
-    "springedge": "^1.0.6",
     "unique-username-generator": "^1.3.0",
     "winston": "^3.12.0"
   }
